Add trivial case for storage directory

+++ src/tagStorageEngine/test.bash
@@ -1,9 +1,14 @@
 printf "Building test\n"
 g++ -Wall -fexceptions -fexpensive-optimizations -O3 -std=c++11  -c tests/test.cpp -o test.o
-g++  -o test test.o  -s  
+g++  -o test test.o  -s
+if [ $? != 0 ]; then
+    exit 1
+fi  
 rm *.o
 printf "Running\n"
+mkdir learned
 ./test
+rm -rf learned
 if [ $? != 0 ]; then
     exit 1
 fi
